# openapi.yaml
openapi: 3.0.3
info:
  title: PescaApp API
  description: Backend RESTful para aplicación de pesca colaborativa
  version: 1.0.0
  contact:
    name: PescaApp Backend
servers:
  - url: https://api.pescaapp.com/v1
    description: Production
  - url: http://localhost:8080/v1
    description: Local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: string
        timestamp:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        photoUrl:
          type: string
          nullable: true
        role:
          type: string
          enum: [user, admin]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserPublic:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        photoUrl:
          type: string
          nullable: true
        stats:
          $ref: '#/components/schemas/UserStats'
    UserStats:
      type: object
      properties:
        spotsCreated:
          type: integer
        commentsCount:
          type: integer
        ratingsCount:
          type: integer
        averageRating:
          type: number
          format: double
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
        photoUrl:
          type: string
          nullable: true
...
    FavoriteSpot:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        region:
          type: string
        rating:
          type: number
          format: double
        addedAt:
          type: string
          format: date-time
...
    SpotStats:
      type: object
      properties:
        spotId:
          type: string
        name:
          type: string
        visits:
          type: integer
        uniqueUsers:
          type: integer
        averageRating:
          type: number
        totalRatings:
          type: integer
        totalComments:
          type: integer
        topSpecies:
          type: array
          items:
            type: object
            properties:
              speciesId:
                type: string
              name:
                type: string
              mentions:
                type: integer
        lastCommentDate:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    UserDetailedStats:
      type: object
      properties:
        userId:
          type: string
        name:
          type: string
        spotsCreated:
          type: integer
        commentsCount:
          type: integer
        ratingsCount:
          type: integer
        averageRating:
          type: number
        favoriteRegions:
          type: array
          items:
            type: string
        favoriteSpecies:
          type: array
          items:
            type: string
        joiningDate:
          type: string
          format: date-time
    PopularSpot:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        region:
          type: string
        averageRating:
          type: number
        totalRatings:
          type: integer
        views:
          type: integer
        totalComments:
          type: integer
  parameters:
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
        maximum: 100
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
  responses:
...
    Forbidden:
      description: Permiso denegado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Listar usuarios
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserPublic'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /users/{id}:
    get:
      tags: [Users]
      summary: Obtener perfil de usuario
      operationId: getUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Perfil de usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Users]
      summary: Actualizar perfil de usuario
      operationId: updateUser
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
  /users/{id}/favorites:
    get:
      tags: [Favorites]
      summary: Listar favoritos del usuario
      operationId: listFavorites
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Lista de favoritos
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/FavoriteSpot'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /users/{id}/favorites/{spotId}:
    post:
      tags: [Favorites]
      summary: Agregar lugar a favoritos
      operationId: addFavorite
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: spotId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Favorito agregado
          content:
            application/json:
              schema:
                type: object
                properties:
                  isFavorite:
                    type: boolean
                  addedAt:
                    type: string
                    format: date-time
        '409':
          description: Ya existe en favoritos
    delete:
      tags: [Favorites]
      summary: Eliminar de favoritos
      operationId: removeFavorite
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: spotId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Eliminado de favoritos
...
  /spots/{id}/stats:
    get:
      tags: [Statistics]
      summary: Estadísticas de un spot
      operationId: getSpotStats
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estadísticas del spot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpotStats'
  /user/{userId}/stats:
    get:
      tags: [Statistics]
      summary: Estadísticas de usuario
      operationId: getUserStats
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Estadísticas del usuario
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDetailedStats'
  /statistics/popular-spots:
    get:
      tags: [Statistics]
      summary: Spots más populares
      operationId: getPopularSpots
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: timeRange
          in: query
          schema:
            type: string
            enum: [week, month, year, all]
        - name: orderBy
          in: query
          schema:
            type: string
            enum: [rating, visits, comments]
            default: rating
      responses:
        '200':
          description: Spots populares
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/PopularSpot'